+++ src/Hacks/RandomSeed.cpp
@@ -1,0 +1,17 @@
+/*#include <Geode/Geode.hpp>
+#include <Geode/modify/PlayLayer.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;
+
+class $modify (PlayLayer)
+{
+    void resetLevel()
+    {
+        PlayLayer::resetLevel();
+
+        srand(69);
+
+        GameToolbox
+    }
+};*/